move verifyFirebaseIdToken to helpers dir
